+++ bitframe-authentication/core/src/commonMain/kotlin/bitframe/authentication/apps/App.kt
@@ -1,0 +1,7 @@
+@file:JsExport
+
+package bitframe.authentication.apps
+
+import kotlin.js.JsExport
+
+data class App(val id: String)