+++ pi-monitor/pi-monitor-client/browser/react/src/jsMain/kotlin/pimonitor/evaluation/business/BusinessList.kt
@@ -3,15 +3,16 @@ package pimonitor.evaluation.business
 import pimonitor.Monitor
 import react.RBuilder
 import reakt.*
+import styled.styledDiv
 
-fun RBuilder.BusinessList(
+internal fun RBuilder.BusinessList(
     data: List<Monitor.Business>
 ) = ReactTable(
     data,
     columns = listOf(
         Column("name") { it.name },
         Column("email") { it.email.value },
-        Column("contact") {  },
+        Column("contact") { it.name },
         RenderColumn("actions") {
             Grid(cols = "1fr 1fr") {
                 ContainedButton("View") {}