+++ pi-monitor/pi-monitor-test/testing/src/jvmMain/kotlin/testing/IntegrationTest.kt
@@ -2,6 +2,7 @@ package testing
 
 import bitframe.service.config.KtorClientConfiguration
 import bitframe.service.config.ServiceConfig
+import testing.pimonitor.APP_ID
 
 actual open class IntegrationTest : ContainerTest() {
     actual val config: ServiceConfig
@@ -11,11 +12,11 @@ actual open class IntegrationTest : ContainerTest() {
         get() = when (mode) {
             TestMode.DEV -> KtorClientConfiguration(
                 url = "http://localhost:8080",
-                appId = "dev-app-1"
+                appId = APP_ID
             )
             TestMode.CI -> KtorClientConfiguration(
                 url = urlUnderTest,
-                appId = "ci-app-1"
+                appId = APP_ID
             )
             TestMode.CD -> TODO()
         }